what the sigma
